# target/generated-sources/openapi/src/main/java/com/nuevospa/taskmanager/AutenticacinApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.16.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nuevospa.taskmanager;

import com.nuevospa.taskmanager.entity.ErrorResponse;
import com.nuevospa.taskmanager.entity.LoginRequest;
import com.nuevospa.taskmanager.entity.LoginResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-13T22:13:05.474109-03:00[America/Santiago]", comments = "Generator version: 7.16.0")
@Validated
@Tag(name = "Autenticación", description = "Endpoints de autenticación y autorización JWT")
public interface AutenticacinApi {

    public static final String PATH_LOGIN = "/auth/login";
    /**
     * POST /auth/login : Autenticar usuario y obtener token JWT
     * Valida las credenciales del usuario y devuelve un token JWT para autenticación
     *
     * @param loginRequest  (required)
     * @return Autenticación exitosa (status code 200)
     *         or Credenciales inválidas (status code 401)
     *         or Solicitud inválida (status code 400)
     */
    @Operation(
        operationId = "login",
        summary = "Autenticar usuario y obtener token JWT",
        description = "Valida las credenciales del usuario y devuelve un token JWT para autenticación",
        tags = { "Autenticación" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Autenticación exitosa", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoginResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Credenciales inválidas", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = AutenticacinApi.PATH_LOGIN,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<LoginResponse> login(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody LoginRequest loginRequest
    );

}