# target/generated-sources/openapi/src/main/java/com/nuevospa/taskmanager/TareasApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.16.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nuevospa.taskmanager;

import com.nuevospa.taskmanager.entity.ErrorResponse;
import com.nuevospa.taskmanager.entity.TareaRequest;
import com.nuevospa.taskmanager.entity.TareaResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-13T22:13:05.474109-03:00[America/Santiago]", comments = "Generator version: 7.16.0")
@Validated
@Tag(name = "Tareas", description = "Operaciones CRUD sobre tareas")
public interface TareasApi {

    public static final String PATH_ACTUALIZAR_TAREA = "/tareas/{id}";
    /**
     * PUT /tareas/{id} : Actualizar una tarea existente
     * Actualiza todos los campos de una tarea específica
     *
     * @param id ID de la tarea a actualizar (required)
     * @param tareaRequest  (required)
     * @return Tarea actualizada exitosamente (status code 200)
     *         or Recurso no encontrado (status code 404)
     *         or Solicitud inválida o datos incorrectos (status code 400)
     *         or Token JWT no válido, expirado o ausente (status code 401)
     */
    @Operation(
        operationId = "actualizarTarea",
        summary = "Actualizar una tarea existente",
        description = "Actualiza todos los campos de una tarea específica",
        tags = { "Tareas" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tarea actualizada exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TareaResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Recurso no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida o datos incorrectos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT no válido, expirado o ausente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = TareasApi.PATH_ACTUALIZAR_TAREA,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<TareaResponse> actualizarTarea(
        @NotNull @Parameter(name = "id", description = "ID de la tarea a actualizar", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "TareaRequest", description = "", required = true) @Valid @RequestBody TareaRequest tareaRequest
    );


    public static final String PATH_CREAR_TAREA = "/tareas";
    /**
     * POST /tareas : Crear una nueva tarea
     * Crea una nueva tarea asociada al usuario autenticado
     *
     * @param tareaRequest  (required)
     * @return Tarea creada exitosamente (status code 201)
     *         or Solicitud inválida o datos incorrectos (status code 400)
     *         or Token JWT no válido, expirado o ausente (status code 401)
     */
    @Operation(
        operationId = "crearTarea",
        summary = "Crear una nueva tarea",
        description = "Crea una nueva tarea asociada al usuario autenticado",
        tags = { "Tareas" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Tarea creada exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TareaResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida o datos incorrectos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT no válido, expirado o ausente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = TareasApi.PATH_CREAR_TAREA,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<TareaResponse> crearTarea(
        @Parameter(name = "TareaRequest", description = "", required = true) @Valid @RequestBody TareaRequest tareaRequest
    );


    public static final String PATH_ELIMINAR_TAREA = "/tareas/{id}";
    /**
     * DELETE /tareas/{id} : Eliminar una tarea
     * Elimina permanentemente una tarea del sistema
     *
     * @param id ID de la tarea a eliminar (required)
     * @return Tarea eliminada exitosamente (status code 204)
     *         or Recurso no encontrado (status code 404)
     *         or Token JWT no válido, expirado o ausente (status code 401)
     */
    @Operation(
        operationId = "eliminarTarea",
        summary = "Eliminar una tarea",
        description = "Elimina permanentemente una tarea del sistema",
        tags = { "Tareas" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Tarea eliminada exitosamente"),
            @ApiResponse(responseCode = "404", description = "Recurso no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT no válido, expirado o ausente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = TareasApi.PATH_ELIMINAR_TAREA,
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> eliminarTarea(
        @NotNull @Parameter(name = "id", description = "ID de la tarea a eliminar", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_LISTAR_TAREAS = "/tareas";
    /**
     * GET /tareas : Listar todas las tareas del usuario
     * Obtiene todas las tareas asociadas al usuario autenticado
     *
     * @return Lista de tareas obtenida exitosamente (status code 200)
     *         or Token JWT no válido, expirado o ausente (status code 401)
     */
    @Operation(
        operationId = "listarTareas",
        summary = "Listar todas las tareas del usuario",
        description = "Obtiene todas las tareas asociadas al usuario autenticado",
        tags = { "Tareas" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de tareas obtenida exitosamente", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TareaResponse.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT no válido, expirado o ausente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TareasApi.PATH_LISTAR_TAREAS,
        produces = { "application/json" }
    )
    
    ResponseEntity<List<TareaResponse>> listarTareas(
        
    );


    public static final String PATH_OBTENER_TAREA = "/tareas/{id}";
    /**
     * GET /tareas/{id} : Obtener una tarea por ID
     * Obtiene los detalles de una tarea específica
     *
     * @param id ID de la tarea a obtener (required)
     * @return Tarea encontrada (status code 200)
     *         or Recurso no encontrado (status code 404)
     *         or Token JWT no válido, expirado o ausente (status code 401)
     */
    @Operation(
        operationId = "obtenerTarea",
        summary = "Obtener una tarea por ID",
        description = "Obtiene los detalles de una tarea específica",
        tags = { "Tareas" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tarea encontrada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TareaResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Recurso no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT no válido, expirado o ausente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TareasApi.PATH_OBTENER_TAREA,
        produces = { "application/json" }
    )
    
    ResponseEntity<TareaResponse> obtenerTarea(
        @NotNull @Parameter(name = "id", description = "ID de la tarea a obtener", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );

}